pyproject.toml
[project]
name = "vinescribe"
version = "0.1.0"
description = ""
authors = [
    {name = "",email = ""}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic (>=2.10.6,<3.0.0)",
    "google-api-python-client (>=2.162.0,<3.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "google-auth-oauthlib (>=1.2.1,<2.0.0)",
    "google-genai (>=1.3.0,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "tenacity (>=9.0.0,<10.0.0)",
    "ratelimit (>=2.2.1,<3.0.0)",
    "pillow (>=11.1.0,<12.0.0)"
]

[tool.poetry]
packages = [
    {include = "vinescribe", from = "src"},
    {include = "services/*", from = "src/vinescribe"}
]


[tool.poetry.group.dev.dependencies]
jupyter = "^1.1.1"
ipykernel = "^6.29.5"
black = "^25.1.0"
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
